chore: move all rest Resource to their own file
