+++ synapse_token_authenticator/token_authenticator.py
@@ -13,12 +13,10 @@
 # You should have received a copy of the GNU Affero General Public License
 # along with this program. If not, see <http://www.gnu.org/licenses/>.
 import base64
-import json
 import logging
 import re
 from collections.abc import Awaitable
 from typing import Callable
-from urllib.parse import urljoin
 
 import synapse
 from jwcrypto import jwk, jwt
@@ -27,12 +25,15 @@ from jwcrypto.jwk import JWKSet
 from synapse.api.errors import HttpResponseException
 from synapse.module_api import ModuleApi
 from synapse.types import UserID
-from twisted.web import resource
 
 from synapse_token_authenticator.auth_headers import basic_auth
 from synapse_token_authenticator.config import TokenAuthenticatorConfig
-from synapse_token_authenticator.utils import (
+from synapse_token_authenticator.rest import (
+    LoginMetadataResource,
     MetadataResource,
+    PublicKeysResource,
+)
+from synapse_token_authenticator.utils import (
     all_list_elems_are_equal_return_the_elem,
     get_oidp_metadata,
     get_path_in_dict,
@@ -73,7 +74,7 @@ class TokenAuthenticator:
 
             self.api.register_web_resource(
                 "/_famedly/login/com.famedly.login.token.oidc",
-                self.LoginMetadataResource(oidc),
+                LoginMetadataResource(oidc),
             )
 
         if (cfg := getattr(self.config, "oauth", None)) is not None:
@@ -100,42 +101,12 @@ class TokenAuthenticator:
             keys = JWKSet()
             keys.add(self.config.epa.enc_jwk)
             self.api.register_web_resource(
-                self.config.epa.enc_jwks_endpoint, self.PublicKeysResource(keys)
+                self.config.epa.enc_jwks_endpoint, PublicKeysResource(keys)
             )
 
             auth_checkers[("com.famedly.login.token.epa", ("token",))] = self.check_epa
 
         self.api.register_password_auth_provider_callbacks(auth_checkers=auth_checkers)
-
-    class LoginMetadataResource(resource.Resource):
-        def __init__(self, oidc_config: object):
-            self.issuer = oidc_config.issuer
-            self.metadata_url = urljoin(
-                oidc_config.issuer, "/.well-known/openid-configuration"
-            )
-            self.organization_id = oidc_config.organization_id
-            self.project_id = oidc_config.project_id
-
-        def render_GET(self, request) -> bytes:
-            request.setHeader(b"content-type", b"application/json")
-            request.setHeader(b"access-control-allow-origin", b"*")
-            return json.dumps(
-                {
-                    "issuer": self.issuer,
-                    "issuer-metadata": self.metadata_url,
-                    "organization-id": self.organization_id,
-                    "project-id": self.project_id,
-                }
-            ).encode("utf-8")
-
-    class PublicKeysResource(resource.Resource):
-        def __init__(self, keys: JWKSet):
-            self.keys = keys.export(private_keys=False).encode("utf-8")
-
-        def render_GET(self, request) -> bytes:
-            request.setHeader(b"content-type", b"application/json")
-            request.setHeader(b"access-control-allow-origin", b"*")
-            return self.keys
 
     async def check_jwt_auth(
         self, username: str, login_type: str, login_dict: "synapse.module_api.JsonDict"
