+++ synapse_token_authenticator/utils.py
@@ -1,9 +1,5 @@
-import json
-from base64 import b64encode
 from typing import Any, List, Optional
 from urllib.parse import urljoin
-
-from twisted.web import resource
 
 
 class OpenIDProviderMetadata:
@@ -63,13 +59,3 @@ def validate_scopes(required_scopes: str | List[str], provided_scopes: str) -> b
         required_scopes = required_scopes.split()
     provided_scopes_list = provided_scopes.split()
     return all(scope in provided_scopes_list for scope in required_scopes)
-
-
-class MetadataResource(resource.Resource):
-    def __init__(self, resource: object) -> None:
-        self.resource = resource
-
-    def render_GET(self, request) -> bytes:
-        request.setHeader(b"content-type", b"application/json")
-        request.setHeader(b"access-control-allow-origin", b"*")
-        return json.dumps(self.resource).encode("utf-8")
